+++ src/velero/providers/__init__.py
@@ -1,0 +1,17 @@
+# Copyright 2025 Canonical Ltd.
+# See LICENSE file for licensing details.
+
+"""Velero Providers module."""
+
+from .azure import AzureStorageConfig, AzureStorageProvider
+from .classes import StorageProviderError, VeleroStorageProvider
+from .s3 import S3StorageConfig, S3StorageProvider
+
+__all__ = [
+    "S3StorageProvider",
+    "AzureStorageProvider",
+    "StorageProviderError",
+    "VeleroStorageProvider",
+    "AzureStorageConfig",
+    "S3StorageConfig",
+]
